fix(mensage-validation): place message
