+++ client/src/components/tela_chatbot/MessageInput.tsx
@@ -1,4 +1,5 @@
-import React from 'react';
+'use client'
+import React, {useState} from 'react';
 import Image from 'next/image';
 import Send from '../../assets/Send.png';
 
@@ -11,22 +12,40 @@ interface MessageInputProps {
 }
 
 const MessageInput: React.FC<MessageInputProps> = ({ message, setMessage, handleKeyPress, handleSendMessage, className }) => {
+  const [error, setError] = useState<string | null>(null);
+
+  const validateAndSend = () => {
+    if (message.trim() === '') { 
+      setError('Por favor, digite uma mensagem ou envie um arquivo.'); 
+    } else {
+      setError(null); 
+      handleSendMessage(); 
+    }
+  };
+
+
   return (
-    <div className={`${className} flex items-center space-x-4`}>
+    <div className={`${className} flex flex-col space-y-2`}>
+      <div className="flex items-center space-x-4">
       <textarea
-      value={message}
-      onChange={(e) => setMessage(e.target.value)}
-      onKeyDown={handleKeyPress}
-      className="w-full p-2 rounded bg-white text-black outline-none focus:ring-0 resize-none"
-      placeholder="Digite sua mensagem..."
-      style={{ height: '100px' }}
+        value={message}
+        onChange={(e) => {
+        setMessage(e.target.value);
+        setError(null); 
+        }}
+        onKeyDown={handleKeyPress}
+        className="w-full p-2 rounded bg-white text-black outline-none focus:ring-0 resize-none"
+        placeholder="Digite sua mensagem..."
+        style={{ height: '100px' }}
       />
       <button
-      onClick={handleSendMessage}
-      className="w-16 h-8 text-white rounded transition-all flex items-center justify-center"
+        onClick={validateAndSend}
+        className="w-16 h-8 text-white rounded transition-all flex items-center justify-center"
       >
-      <Image src={Send} alt="Send Icon" className="w-8 h-8" />
+        <Image src={Send} alt="Send Icon" className="w-8 h-8" />
       </button>
+      </div>
+      {error && <p className="text-red-500 text-sm font-semibold">{error}</p>}
     </div>
   );
 };
